+++ NEWS.md
@@ -1,3 +1,9 @@
+## 1.3.0 unpublished
+ * CQLPG-6: Replace regexp by LIKE for <> and == for performance
+ * CQLPG-10: Use unaccent instead of regexp for performance
+## 1.2.1 2017-11-14
+ * Replace CASE jsonb_typeof(field) ... by an AND/OR expression (CQLPG-8)
+ * For numeric values use -> not ->> to make use of the index. (CQLPG-7 partially)
 ## 1.2.0 2017-07-20
  * Implement multi-field queries (FOLIO-727)
 ## 1.1.0 2017-05-10